+++ src/test/java/roomescape/acceptance/ReservationAcceptanceTest.java
@@ -399,52 +399,4 @@ class ReservationAcceptanceTest extends AcceptanceTest {
                 })
         );
     }
-
-    @DisplayName("본인의 모든 예약과 예약 대기를 조회한다.")
-    @TestFactory
-    Stream<DynamicTest> showAllOwnReservationsAndWaitings() {
-        return Stream.of(
-                DynamicTest.dynamicTest("guest1이 새로운 예약을 추가한다.", () -> {
-                    RestAssured.given().log().all()
-                            .contentType(ContentType.JSON)
-                            .cookie("token", guest1Token)
-                            .body(new ReservationRequest(date, timeId, themeId))
-                            .when().post("/reservations")
-                            .then().log().all()
-                            .assertThat().statusCode(201);
-                }),
-                DynamicTest.dynamicTest("guest1이 본인의 예약 내역을 조회하면 1개 내역이 조회된다.", () -> {
-                    RestAssured.given().log().all()
-                            .cookie("token", guest1Token)
-                            .when().get("/members/reservations")
-                            .then().log().all()
-                            .assertThat().statusCode(200).body("size()", is(1));
-                }),
-                DynamicTest.dynamicTest("guest2가 새로운 예약을 추가한다.", () -> {
-                    RestAssured.given().log().all()
-                            .contentType(ContentType.JSON)
-                            .cookie("token", guest1Token)
-                            .body(new ReservationRequest(LocalDate.MAX, timeId, themeId))
-                            .when().post("/reservations")
-                            .then().log().all()
-                            .assertThat().statusCode(201);
-                }),
-                DynamicTest.dynamicTest("guest1이 guest2와 동일한 테마와 일정으로 예약을 요청하고, 1번째 예약 대기 상태로 생성된다.", () -> {
-                    RestAssured.given().log().all()
-                            .contentType(ContentType.JSON)
-                            .cookie("token", guest2Token)
-                            .body(new ReservationRequest(LocalDate.MAX, timeId, themeId))
-                            .when().post("/reservations")
-                            .then().log().all()
-                            .assertThat().statusCode(201).body("status", is("1번째 대기"));
-                }),
-                DynamicTest.dynamicTest("guest1이 본인의 예약 내역을 조회하면 2개 내역이 조회된다.", () -> {
-                    RestAssured.given().log().all()
-                            .cookie("token", guest1Token)
-                            .when().get("/members/reservations")
-                            .then().log().all()
-                            .assertThat().statusCode(200).body("size()", is(2));
-                })
-        );
-    }
 }
